// src/app/model/near-region-territory.ts
export interface NearRegionTerritory {
  id?: string;
  near_RegionCode?: string;
  near_RegionDescription?: string;
  central_regionCode?: string;
  central_regionDescription?: string;
  priority?: number;
  sysCreatedBy?: string;
  sysCreatedDate?: Date;
  sysUpdatedBy?: string;
  sysUpdatedDate?: Date;
}